Move to int tests.
