+++ crates/cli/src/commands/bin.rs
@@ -47,37 +47,3 @@ pub async fn bin(tool_type: &BinTools) -> Result<(), Box<dyn std::error::Error>>
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::helpers::create_test_command;
-    use predicates::prelude::*;
-
-    #[test]
-    fn invalid_tool() {
-        let assert = create_test_command("base")
-            .arg("bin")
-            .arg("unknown")
-            .assert();
-
-        assert
-            .failure()
-            .code(2)
-            .stdout("")
-            .stderr(predicate::str::contains("\"unknown\" isn\'t a valid value"));
-    }
-
-    #[test]
-    fn not_configured() {
-        let assert = create_test_command("base").arg("bin").arg("yarn").assert();
-
-        assert.failure().code(1).stdout("");
-    }
-
-    #[test]
-    fn not_installed() {
-        let assert = create_test_command("base").arg("bin").arg("node").assert();
-
-        assert.failure().code(2).stdout("");
-    }
-}
