+++ crates/cli/src/commands/setup.rs
@@ -8,19 +8,3 @@ pub async fn setup() -> Result<(), Box<dyn std::error::Error>> {
 
     Ok(())
 }
-
-// #[cfg(test)]
-// mod tests {
-//     use crate::helpers::create_test_command;
-
-//     #[test]
-//     fn installs() {
-//         let assert = create_test_command("base")
-//             .arg("--log-level")
-//             .arg("trace")
-//             .arg("setup")
-//             .assert();
-
-//         assert.success().code(0);
-//     }
-// }
