+++ crates/cli/src/commands/teardown.rs
@@ -7,15 +7,3 @@ pub async fn teardown() -> Result<(), Box<dyn std::error::Error>> {
 
     Ok(())
 }
-
-// #[cfg(test)]
-// mod tests {
-//     use crate::helpers::create_test_command;
-
-//     #[test]
-//     fn uninstalls() {
-//         let assert = create_test_command("base").arg("teardown").assert();
-
-//         assert.success().code(0);
-//     }
-// }
