+++ tests/fixtures/cases/node-throw-error/project.yml
@@ -1,0 +1,4 @@
+tasks:
+  test:
+    command: node
+    args: ./test.js
