+++ tests/fixtures/cases/node-throw-error/test.js
@@ -1,0 +1,1 @@
+throw new Error('Oops');
