+++ tests/fixtures/cases/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "test-cases"
+}
